src/front-end/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
jinja2==3.1.3
httpx==0.25.0 

bcrypt==4.0.1